parser/requirements.txt
requests
asyncpg
aiohttp
python-dateutil